refactor(core): Use model_name instead of ticket_comment_model for url kwarg
ref: #1128
