+++ app/core/tests/unit/ticket_comment_task/test_unit_ticket_comment_task_viewset.py
@@ -73,7 +73,7 @@ class ViewsetTestCases(
 
 
         view_set.kwargs = {
-            'ticket_comment_model': 'task',
+            'model_name': 'ticketcommenttask',
             'ticket_id': ticket.id
         }
