+++ app/tests/fixtures/model_ticketcommentaction.py
@@ -1,7 +1,7 @@
 import pytest
 
 from core.models.ticket_comment_action import TicketCommentAction
-from core.serializers.ticketcommentbase_action import (
+from core.serializers.ticketcommentbase_ticketcommentaction import (
     BaseSerializer,
     ModelSerializer,
     ViewSerializer,
